TransfusionService: omit deprecated cPar::longValue() call

+++ src/artery/transfusion/TransfusionService.cc
@@ -43,7 +43,7 @@ void TransfusionService::initialize()
     if (ec) {
         throw cRuntimeError("remote IP address is invalid: %s", par("remote_ip").stringValue());
     }
-    unsigned remote_port = par("remote_port").longValue();
+    unsigned remote_port = par("remote_port");
     endpoint.address(remote_ip);
     endpoint.port(remote_port);
